+++ packages/tables/resources/views/cells/boolean.blade.php
@@ -1,5 +1,3 @@
-@php
-$classes = $column->getValue($record) ? 'text-primary-600' : 'text-danger-700';
-@endphp
-
-@include('tables::cells.icon')
+@include('tables::cells.icon', [
+    'classes' => $column->getValue($record) ? 'text-primary-600' : 'text-danger-700',
+])
