+++ packages/tables/resources/views/cells/icon.blade.php
@@ -1,24 +1,36 @@
 @php
-$iconToShow = null;
+    $iconToShow = null;
 
-foreach ($column->options as $icon => $callback) {
-    if (! $callback($column->getValue($record))) {
-        continue;
+    foreach ($column->options as $icon => $callback) {
+        if (! $callback($column->getValue($record))) {
+            continue;
+        }
+
+        $iconToShow = $icon;
+
+        break;
     }
-
-    $iconToShow = $icon;
-
-    break;
-}
 @endphp
 
-@if($column->action)
-    <button
-        wire:click="{{ $column->action }}('{{ $record->getKey() }}')"
-        type="button"
-    >
-        <x-dynamic-component :component="$iconToShow" class="{{ $classes ?? '' }} w-6 h-6" />
-    </button>
-@else
-    <x-dynamic-component :component="$iconToShow" class="{{ $classes ?? '' }} w-6 h-6" />
+@if ($iconToShow)
+    @if ($column->action)
+        <button
+            wire:click="{{ $column->action }}('{{ $record->getKey() }}')"
+            type="button"
+        >
+            <x-dynamic-component :component="$iconToShow" class="{{ $classes ?? null }} w-6 h-6" />
+        </button>
+    @elseif ($column->url)
+        <a
+            href="{{ $column->getUrl($record) }}"
+            @if ($column->shouldOpenUrlInNewTab)
+            target="_blank"
+            rel="noopener noreferrer"
+            @endif
+        >
+            <x-dynamic-component :component="$iconToShow" class="{{ $classes ?? null }} w-6 h-6" />
+        </a>
+    @else
+        <x-dynamic-component :component="$iconToShow" class="{{ $classes ?? null }} w-6 h-6" />
+    @endif
 @endif
