+++ packages/tables/src/Columns/Text.php
@@ -6,15 +6,12 @@ use Carbon\Carbon;
 
 class Text extends Column
 {
-    use Concerns\HasAction;
+    use Concerns\CanCallAction;
+    use Concerns\CanOpenUrl;
 
     public $default;
 
     public $formatUsing;
-
-    public $shouldOpenUrlInNewTab = false;
-
-    public $url;
 
     public function currency($symbol = '$', $decimalSeparator = '.', $thousandsSeparator = ',')
     {
@@ -59,17 +56,6 @@ class Text extends Column
         return $this;
     }
 
-    public function getUrl($record)
-    {
-        if (is_callable($this->url)) {
-            $callback = $this->url;
-
-            return $callback($record);
-        }
-
-        return $this->url;
-    }
-
     public function getValue($record, $attribute = null)
     {
         $value = parent::getValue($record, $attribute);
@@ -87,13 +73,6 @@ class Text extends Column
         return $value;
     }
 
-    public function openUrlInNewTab()
-    {
-        $this->shouldOpenUrlInNewTab = true;
-
-        return $this;
-    }
-
     public function options($options)
     {
         $this->formatUsing = function ($value) use ($options) {
@@ -103,14 +82,6 @@ class Text extends Column
 
             return $value;
         };
-
-        return $this;
-    }
-
-    public function url($url, $shouldOpenInNewTab = false)
-    {
-        $this->url = $url;
-        if ($shouldOpenInNewTab) $this->openUrlInNewTab();
 
         return $this;
     }
